+++ Source/Game.h
@@ -12,6 +12,7 @@ private:
 	SDL_Event event = {};
 	std::unique_ptr<UI> ui = nullptr;
 
+	TagUISection tagSec;
 
 public:
 	Game();