+++ CHANGELOG.md
@@ -93,3 +93,4 @@ was removed or the new one was added to folder
 - New tool Path Maker make for simple and fast custom path creation
 - Retiveing elemnts from UI is 2 times faster
 - Removing ui elements is safer and no longer requires element type
+- New UI Helper container TagUISection designed to vastly simplyfy managing large interfaces